requirements.txt
Django==1.8.6
djangorestframework==3.3.1
simplejson==3.8.1
MySQL-python==1.2.5